Split Inspect rows and shared UI widgets into folders.
