+++ packages/devtools/src/ui/ShiroLogo.tsx
