+++ packages/devtools/src/ui/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 import { Moon, Sun } from "lucide-react";
 import type { CSSProperties } from "react";
 import { IconButton } from "./IconButton";
-import { useStyles, type ColorMode, type ThemeTokens } from "./theme";
+import { useStyles, type ColorMode, type ThemeTokens } from "../theme";
 
 const OPTIONS: { id: ColorMode; label: string; icon: typeof Sun }[] = [
   { id: "light", label: "Light", icon: Sun },
